+++ src/app/exercise/exercises/utility/answer-layouts/scale-layout.ts
